requirements.txt
Flask
gunicorn
flask_session
selenium
beautifulsoup4
psycopg2-binary
werkzeug
requests
numpy
